+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [2.1.0]
+ - Add double quoted tokens support to HUD Animations Language Server
  - Add features to Popfile Language Server
    - Add Colours provider
    - Add Definition/References provider